d_a_bigelf: next_msgStatus 95.2->100 (ItemMax accessors 0x70/0x6F real value fix; retail dispatch: 0x2EE9/2EEA->default 0x10, 2EEB/2EEC/2EEF/2EF0->inc, 2EF1->0x2EEE), talk 98.9 (s32 next_msgStatus ret kills clrlwi). Fresh .o postdates.

+++ include/d/actor/d_a_bigelf.h
@@ -73,7 +73,7 @@ public:
     void makeFa1();
     void setAnm(signed char);
     void setAnmStatus();
-    u8 next_msgStatus(unsigned long*);
+    s32 next_msgStatus(unsigned long*);
     u32 getMsg();
     void msgPushButton();
     void msgAnm(unsigned char);
